:recycle: [Refactor]: DTO 클래스 구조 변경
